test: 기능 분리에 따른 테스트 코드 수정
