+++ src/test/java/study/carracing/domain/RacingTest.java
@@ -6,42 +6,29 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class RacingTest {
+    private RacingCars racingCars = new RacingCars();
+    private Racing racing = new Racing();
 
     @Test
-    @DisplayName("문자열의 콤마(,)를 기준으로 split한 배열의 length만큼 자동차가 생성된다.")
-    void split_car_name() {
-        RacingCars racingCars = new RacingCars();
-
-        racingCars.addCar("car1,car2,car3".split(","));
-
-        assertThat(racingCars.getCars().size()).isEqualTo(3);
-    }
-
-    @Test
-    @DisplayName("꺼내온 자동차들의 값을 변경하면 UnsupportedOperationException예외가 발생한다.")
-    void not_allow_to_modify_racingCars() {
-        Racing racing = new Racing();
+    @DisplayName("게임 히스토리 값을 변경하면 UnsupportedOperationException예외가 발생한다.")
+    void not_allow_to_modify_history() {
         racing.start("car1,car2,car3".split(","), 3);
-        List<Car> racingCars = racing.getRacingCars();
+        List<String> scores = racing.getHistory().getScores();
 
-        assertThatThrownBy(() -> racingCars.add(new Car(new Name("pobi"))))
+        assertThatThrownBy(() -> scores.add("test score"))
                 .isInstanceOf(UnsupportedOperationException.class);
-
     }
 
     @Test
     @DisplayName("꺼내온 우승자 리스트의 값을 변경하면 UnsupportedOperationException예외가 발생한다.")
     void not_allow_to_modify_winners() {
-        Racing racing = new Racing();
         racing.start("car1,car2,car3".split(","), 3);
         List<String> winners = racing.winners();
 
         assertThatThrownBy(() -> winners.add("new winner"))
                 .isInstanceOf(UnsupportedOperationException.class);
-
     }
 }
